requirements.txt
vinted-scraper
python-telegram-bot
httpx